Added Units Tests (JUnit)
Tests are for Parser (test all public methods (but not exhaustively)) and for TaskList (only for the size() method)
