+++ src/main/java/duke/commands/ListCommand.java
@@ -1,8 +1,10 @@
 package duke.commands;
 
-import duke.exceptions.*;
-import duke.tasks.*;
-import duke.utils.*;
+import duke.exceptions.DukeException;
+import duke.tasks.Task;
+import duke.utils.TaskList;
+import duke.utils.Storage;
+import duke.utils.Ui;
 
 import java.util.ArrayList;
