+++ src/main/java/duke/utils/Storage.java
@@ -2,7 +2,10 @@ package duke.utils;
 
 import duke.exceptions.DukeException;
 import duke.Duke;
-import duke.tasks.*;
+import duke.tasks.Task;
+import duke.tasks.ToDo;
+import duke.tasks.Event;
+import duke.tasks.Deadline;
 
 import java.io.File;
 import java.io.FileWriter;
