+++ src/main/java/duke/utils/TaskList.java
@@ -1,6 +1,9 @@
 package duke.utils;
 
-import duke.tasks.*;
+import duke.tasks.Task;
+import duke.tasks.ToDo;
+import duke.tasks.Event;
+import duke.tasks.Deadline;
 import duke.exceptions.DukeException;
 
 import java.util.ArrayList;
